1.2.5   * データテキスト/検索機能を追加   * データファイル/検索機能を追加   * ピース/更新用複製機能を追加   * ピース/プレビュー機能を追加   * ピース/検索機能を追加   * レイアウト/プレビュー機能を追加   * レイアウト/検索機能を追加   * スタイルシート/権限機能を追加   * ディレクトリ/携帯プレビュー機能を追加   * 記事/新着一覧ピース/一覧リンク文字オプションを追加   * 新着ポータル/フィード出力機能を追加   * サンプルデータ/スマートフォンサイト紹介ページを追加   * ふりがな/特定文字列でふりがなが表示されない不具合を修正   * 記事/プレビュー時に画像が表示されない不具合を修正   * 記事/更新用複製を下書き保存した際に公開中の添付ファイルが削除される不具合を 修正   * 記事/更新用複製/編集権限がなくてもメニューが表示されている不具合を修正   * 記事/分野別一覧/新着記事がない場合に記事一覧が表示されない不具合を修正   * アンケート/フォーム画面のプレビューが表示できない不具合を修正
1.2.5
  CREATE TABLE cms_stylesheets (
    id int(11) NOT NULL AUTO_INCREMENT,
    concept_id int(11) DEFAULT NULL,
    site_id int(11) DEFAULT NULL,
    created_at datetime DEFAULT NULL,
    updated_at datetime DEFAULT NULL,
    path varchar(255) DEFAULT NULL,
    PRIMARY KEY (id)
  );

+++ app/controllers/cms/admin/stylesheets_controller.rb
@@ -4,21 +4,17 @@ class Cms::Admin::StylesheetsController < Cms::Controller::Admin::Base
   
   def pre_dispatch
     return error_auth unless Core.user.has_auth?(:designer)
-    @root      = "#{Rails.root}/public/_common/themes"
     @path      = params[:path].join('/').to_s.force_encoding('utf-8')
-    @full_path = "#{@root}/#{@path}"
-    @base_uri  = ["#{Rails.root}/public/", "/"]
+    @item = Cms::Stylesheet.new_by_path(@path)
     
-    unless ::File.exist?(@full_path)
+    unless ::File.exist?(@item.upload_path)
       return http_error(404) if flash[:notice]
-      flash[:notice] = "指定されたパスは存在しません。（ #{@full_path.gsub(@base_uri[0], @base_uri[1])} ）"
+      flash[:notice] = "指定されたパスは存在しません。（ #{@item.upload_path} ）"
       redirect_to(cms_stylesheets_path(''))
     end
   end
   
   def index
-    @item = Cms::Stylesheet.find(@full_path, :root => @root, :base_uri => @base_uri)
-    
     return show    if params[:do] == 'show'
     return new     if params[:do] == 'new'
     return edit    if params[:do] == 'edit'
@@ -39,42 +35,28 @@ class Cms::Admin::StylesheetsController < Cms::Controller::Admin::Base
   end
   
   def show
+    #return error_auth unless @item.readable?
+    
     @item.read_body
     render :action => :show
   end
   
   def new
+    return error_auth unless @item.creatable?
+    
     render :action => :new
   end
   
   def edit
+    return error_auth unless @item.editable?
+    
     @item.read_body
     render :action => :edit
   end
   
-  def rename
-    if request.put?
-      if @item.rename(params[:item][:name])
-        flash[:notice] = '更新処理が完了しました。'
-        location = cms_stylesheets_path(::File.dirname(@path))
-        return redirect_to(location)
-      end
-    end
-    render :action => :rename
-  end
-  
-  def move
-    if request.put?
-      if @item.move(params[:item][:path])
-        flash[:notice] = '更新処理が完了しました。'
-        location = cms_stylesheets_path(::File.dirname(@path))
-        return redirect_to(location)
-      end
-    end
-    render :action => :move
-  end
-  
   def create
+    return error_auth unless @item.creatable?
+      
     if params[:create_directory]
       if @item.create_directory(params[:item][:new_directory])
         flash[:notice] = 'ディレクトリを作成しました。'
@@ -94,12 +76,42 @@ class Cms::Admin::StylesheetsController < Cms::Controller::Admin::Base
     return false
   end
   
+  def rename
+    return error_auth unless @item.editable?
+    
+    if request.put?
+      @item.concept_id = params[:item][:concept_id]
+      @item.site_id    = Core.site.id if @item.concept_id 
+      
+      if @item.rename(params[:item][:name])
+        flash[:notice] = '更新処理が完了しました。'
+        location = cms_stylesheets_path(::File.dirname(@path))
+        return redirect_to(location)
+      end
+    end
+    render :action => :rename
+  end
+  
+  def move
+    return error_auth unless @item.editable?
+    
+    if request.put?
+      if @item.move(params[:item][:path])
+        flash[:notice] = '更新処理が完了しました。'
+        location = cms_stylesheets_path(::File.dirname(@path))
+        return redirect_to(location)
+      end
+    end
+    render :action => :move
+  end
+  
   def update
+    return error_auth unless @item.editable?
+    
     @item.body = params[:item][:body]
     
-    if @item.valid? && @item.save
+    if @item.valid? && @item.update_file
       flash[:notice] = '更新処理が完了しました。'
-      #location = cms_stylesheets_path(::File.dirname(@path))
       location = cms_stylesheets_path(@path, :do => 'edit')
       return redirect_to(location)
     end
@@ -107,6 +119,8 @@ class Cms::Admin::StylesheetsController < Cms::Controller::Admin::Base
   end
   
   def destroy
+    return error_auth unless @item.deletable?
+    
     if @item.destroy
       flash[:notice] = "削除処理が完了しました。"
     else
